pyproject.toml
[project]
name = "langchain-gpt-oss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.30",
    "load-dotenv>=0.1.0",
]
